requirements.txt
streamlit>=1.30.0
deepface>=0.0.83
opencv-python-headless>=4.5.5
tensorflow>=2.9.1,<2.12
pandas>=1.3.0
Pillow>=9.0.0
